+++ frontend/app/src/main/java/com/example/frontend/usecase/LoginUseCase.kt
@@ -68,6 +68,7 @@ class LoginUseCase(
                 Toast.makeText(context, result.value, Toast.LENGTH_LONG).show()
             }
         })
+
     }
